QPID-1803 : Fixed potential NPE in FileUtils.delete merged from trunk r764076
git-svn-id: https://svn.apache.org/repos/asf/qpid/branches/0.5-release@764116 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpid/util/FileUtils.java
@@ -250,9 +250,17 @@ public class FileUtils
          {
              if (recursive)
              {
-                 for (File subFile : file.listFiles())
+                 File[] files = file.listFiles();
+
+                 // This can occur if the file is deleted outside the JVM
+                 if (files == null)
                  {
-                     success = delete(subFile, true) && success;
+                     return false;
+                 }
+
+                 for (int i = 0; i < files.length; i++)
+                 {
+                     success = delete(files[i], true) && success;
                  }
 
                  return success && file.delete();
